CHORE: remove old module
